Refactor user import functionality; add cache invalidation and improve error handling in user views
