+++ src/frontend/admin/templates/role/role_form.html
@@ -1,4 +1,3 @@
-{% set error = error | default({}) %}
 {% set create = request.path[-1:] == '0' | default('true') %}
 {% from "macros/table.html" import table with context %}
 {% set form_title = 'Role' %}
@@ -8,8 +7,7 @@
         'id': '',
         'name': '',
         'description': '',
-        'permissions': [],
-        '_errors': {}
+        'permissions': []
     })
 %}
 
@@ -19,8 +17,8 @@
     <h1 class="text-3xl font-bold mb-5 w-4/5 mx-auto">{% if create %}Create{% else %}Edit{% endif %} {{ form_title }}</h1>
 
     <form id="role-form" {% if create %}hx-post="/frontend/roles"{% else %} hx-put="/frontend/roles/{{ role.id }}" {% endif %} hx-swap="innerHTML"  hx-target="#form-container" hx-target-error="#form-container" class="w-4/5 mx-auto">
-        {% if error.globalerror %}
-        <p class="text-3xl font-bold mb-4 error-msg">{{ error.globalerror }}</p>
+        {% if error %}
+        <p class="text-3xl font-bold mb-4 error-msg">{{ error }}</p>
         {% endif %}
 
         <label class="floating-label mb-5 w-3/5">
