+++ src/frontend/admin/templates/user/user_form.html
@@ -1,4 +1,3 @@
-{% set error = error | default({}) %}
 {% set create = request.path[-1:] == '0' | default('true') %}
 {% set form_title = 'User' %}
 
@@ -8,14 +7,18 @@
         'username': '',
         'name': '',
         'roles': [],
-        'organization': '',
-        '_errors': {}
+        'organization': ''
     })
 %}
 
 {% set form_error = form_error | default({}) %}
 {% set roles = roles | default([]) %}
 {% set organizations = organizations | default([]) %}
+{% if user.roles[0] is integer %}
+  {% set user_roles = user.roles %}
+{% else %}
+  {% set user_roles = user.roles | map(attribute='id') | list %}
+{% endif %}
 <div class="bg-base-100 rounded-sm p-4 mt-4 wx-4/5 mx-auto">
     <div class="w-4/5 mx-auto">
     {% if create -%}
@@ -26,19 +29,29 @@
     </div>
 
     <form id="user-form" class="w-4/5 mx-auto" {% if create %}hx-post="/frontend/users"{% else %}hx-put="/frontend/users/{{ user.id }}"{% endif %} hx-swap="innerHTML"  hx-target="#form-container" hx-target-error="#form-container">
-        {% if error.globalerror -%}
-        <p class="text-3xl font-bold mb-4 error-msg">{{ error.globalerror }}</p>
+        {% if error -%}
+        <p class="text-3xl font-bold mb-4 error-msg">{{ error }}</p>
         {% endif -%}
 
-        <label class="floating-label mb-5 w-3/5">
-            <span>Username</span>
-            <input id="username" type="text" placeholder="Username" class="input w-3/5 validator {% if form_error.username %}input-error{% endif %}" name="username" required value="{{ user.username }}"/>
-        </label>
+        <div class="mb-5 w-3/5 flex flex-row">
+          <label class="floating-label w-3/5 flex-none">
+              <span>Username</span>
+              <input id="username" type="text" placeholder="Username" class="input w-full validator {% if form_error.username %}input-error{% endif %}" name="username" required value="{{ user.username }}"/>
+          </label>
+          {% if form_error.username %}
+            <div role="alert" class="alert alert-error alert-soft w-2/5 flex-initial">{{ form_error.username }}</div>
+          {% endif %}
+        </div>
 
-        <label class="floating-label mb-5 w-3/5">
+        <div class="mb-5 w-3/5 flex flex-row">
+        <label class="floating-label w-3/5 flex-none">
             <span>Name</span>
-            <input type="text" placeholder="Name" class="input w-3/5 validator" name="name" required value="{{ user.name }}" id="userform_name" />
+            <input type="text" placeholder="Name" class="input w-full validator {% if form_error.name %}input-error{% endif %}" name="name" required value="{{ user.name }}" id="userform_name" />
         </label>
+        {% if form_error.name %}
+        <div role="alert" class="alert alert-error alert-soft w-2/5 flex-initial">{{ form_error.username }}</div>
+        {% endif %}
+        </div>
 
         <div x-data="{ showPassword: false }" class="join gap-1 w-3/5">
           <label class="floating-label mb-5 input validator w-3/5">
@@ -63,7 +76,7 @@
         </label>
         <select name="roles[]" multiple required type="number" id="user-role-select">
             {% for role in roles %}
-                <option value="{{ role.id }}" {% if role.id in user.roles %}selected{% endif %}>{{ role.name }} - {{ role.description | default(' ', true) | truncate(50, true) }}</option>
+                <option value="{{ role.id }}" {% if role.id in user_roles %}selected{% endif %}>{{ role.name }} - {{ role.description | default(' ', true) | truncate(50, true) }}</option>
             {% endfor %}
         </select>
         <p class="validator-hint mb-5">Required</p>
