+++ src/frontend/admin/templates/user/users_table_header.html
@@ -8,7 +8,7 @@
   </div>
   {{ table_search_bar('users-table') }}
   <div class="grow-0">
-      <div class="join gap-1">
+      <div class="join gap-1" hx-target="#form-container">
       {{ import_button(url_for('admin.import_users')) }}
       {{ export_button(url_for('admin.export_users')) }}
       </div>
